Slack testing and cleanup (new) slack message testing routines (new) random NotifyPackage routine for testing emails & slack (new) exposed rate manager objects so can get config and metrics (fix) relocated testing routines into Notify class
