+++ server/records/notify/notifyEmail.ts
@@ -3,7 +3,6 @@ import * as UTIL from '../utils/utils';
 import { ENVIRONMENT_TYPE } from '../../config';
 import { NotifyPackage, NotifyType, getMessagePrefixByType, getMessageIconUrlByType } from './notifyShared';
 import { RateManager, RateManagerConfig, RateManagerResult } from '../utils/rateManager';
-import { Logger as LOG, LogSection } from '../logger/log';
 
 // declaring this empty for branding/clarity since it is used
 // for instances that are not related to the RateManager
@@ -27,6 +26,9 @@ export class NotifyEmail {
     public static isActive(): boolean {
         // we're initialized if we have a logger running
         return (NotifyEmail.rateManager!=null);
+    }
+    public static getRateManager(): RateManager<EmailEntry> | null {
+        return NotifyEmail.rateManager;
     }
     public static configure(env?: ENVIRONMENT_TYPE, targetRate?: number, burstRate?: number, burstThreshold?: number): EmailResult {
         // we allow for re-assigning configuration options even if already running
@@ -346,126 +348,6 @@ export class NotifyEmail {
         } catch (error) {
             return { success: false, message: 'failed to send message', data: { error: UTIL.getErrorString(error) } };
         }
-    }
-    //#endregion
-
-    //#region TESTING
-    private static randomNotifyPackage(index: number): NotifyPackage {
-
-        const getRandomInt= (min: number, max: number): number => {
-            return Math.floor(Math.random() * (max - min + 1)) + min;
-        };
-
-        // predefined options for each type
-        const messages = {
-            [NotifyType.SYSTEM_ERROR]: [
-                'A critical system error has occurred. Immediate attention required.',
-                'System failure detected in the main service pipeline.',
-                'Unexpected system downtime. Engineers have been notified.',
-                'Database connection failure. Systems are currently offline.',
-                'Critical resource unavailable. Check service status immediately.'
-            ],
-            [NotifyType.SYSTEM_NOTICE]: [
-                'System maintenance is scheduled for tomorrow at 3 AM.',
-                'System updates will be applied this weekend. Expect downtime.',
-                'Routine server checks are complete. All systems are operational.',
-                'A new feature has been deployed successfully. No action required.',
-                'Configuration changes have been made. Please verify system performance.'
-            ],
-            [NotifyType.JOB_FAILED]: [
-                'The job you submitted has failed due to a processing error.',
-                'Job processing could not be completed. Please check inputs.',
-                'Job failure detected during data validation.',
-                'Your job did not complete successfully. Logs have been updated.',
-                'Unexpected failure during job execution. Please retry.'
-            ],
-            [NotifyType.JOB_PASSED]: [
-                'Your job has successfully completed.',
-                'Job execution finished without errors. Data is ready for review.',
-                'Processing complete. Results have been stored and are available.',
-                'The job you submitted completed successfully. No further action needed.',
-                'All tasks finished without issues. You can access the output.'
-            ],
-            [NotifyType.JOB_STARTED]: [
-                'A new job has started processing.',
-                'Job execution has begun. You will be notified when it completes.',
-                'Your job is now in progress.',
-                'The system has started processing your job. Please wait for further updates.',
-                'Processing of the submitted job has commenced.'
-            ],
-            [NotifyType.SECURITY_NOTICE]: [
-                'A security alert has been triggered on your account.',
-                'Unusual activity detected. Please verify your login credentials.',
-                'Security protocols have been activated due to suspicious access.',
-                'System lockdown initiated due to a security breach.',
-                'Unauthorized login attempt detected. Immediate action required.'
-            ]
-        };
-        const details = [
-            { url: 'https://example.com/job-details', label: 'Job Details' },
-            { url: 'https://example.com/error-log', label: 'Error Log' },
-            { url: 'https://example.com/security-alert', label: 'Security Alert' },
-            { url: 'https://example.com/system-status', label: 'System Status' }
-        ];
-        const detailsMessages = [
-            'Check the logs for further details.',
-            'Please contact support if you need further assistance.',
-            'More information can be found in the error log.',
-            'Refer to the system documentation for troubleshooting.',
-            'Visit the provided link for additional details.',
-            'The issue has been escalated to the dev team.',
-            'System updates will follow shortly. Stay tuned.'
-        ];
-        const types = [
-            NotifyType.SYSTEM_ERROR,
-            NotifyType.SYSTEM_NOTICE,
-            NotifyType.JOB_FAILED,
-            NotifyType.JOB_PASSED,
-            NotifyType.JOB_STARTED,
-            NotifyType.SECURITY_NOTICE
-        ];
-
-        // our date range for the entry
-        const notifyType = types[getRandomInt(0, types.length - 1)];
-        const currentDate = new Date();
-        const startDate = new Date(currentDate.getTime() - getRandomInt(1000000, 100000000)); // random past start date
-        const endDate = Math.random() > 0.5 ? new Date(startDate.getTime() + getRandomInt(1000000, 100000000)) : undefined;
-
-        // structure our return
-        return {
-            type: notifyType,
-            message: `${index}: ${messages[notifyType][getRandomInt(0, 4)]}`,
-            startDate,
-            endDate,
-            detailsLink: Math.random() > 0.5 ? details[getRandomInt(0, details.length - 1)] : undefined,
-            detailsMessage: detailsMessages[getRandomInt(0, detailsMessages.length - 1)],
-            sendTo: ['emaslowski@quotient-inc.com']
-        };
-    }
-    public static async testEmails(numEmails: number): Promise<EmailResult> {
-
-        const hasRateManager: boolean = (NotifyEmail.rateManager) ? (NotifyEmail.rateManager!=null) : false;
-        const config: RateManagerConfig<EmailEntry> | null = NotifyEmail.rateManager?.getConfig() ?? null;
-
-        // create our profiler
-        // we use a random string in case another test or profile is run to avoid collisisons
-        const profileKey: string = `EmailTest_${Math.random().toString(36).substring(2, 6)}`;
-        await LOG.profile(profileKey, LogSection.eSYS, `Email test: ${new Date().toLocaleString()}`, {
-            numEmails,
-            rateManager: hasRateManager,
-            ...(hasRateManager === true && config && {
-                config: (({ onPost: _onPost, ...rest }) => rest)(config)  // Exclude onPost
-            })
-        },'NotifyEmail.test');
-
-        // test our emails
-        for(let i=0; i<numEmails; ++i)
-            await NotifyEmail.sendMessage(NotifyEmail.randomNotifyPackage(i));
-
-        // close our profiler and return results
-        const metrics = NotifyEmail.rateManager?.getMetrics();
-        const result = await LOG.profileEnd(profileKey);
-        return { success: true, message: `finished testing ${numEmails} emails.`, data: { message: result.message, maxRate: metrics?.rates.max, avgRate: metrics?.rates.average } };
     }
     //#endregion
 }
